fix(react-router): extracted hash scroll handling into react-core
