+++ packages/solid-router/src/Transitioner.tsx
@@ -1,5 +1,9 @@
 import * as Solid from 'solid-js'
-import { getLocationChangeInfo, trimPathRight } from '@tanstack/router-core'
+import {
+  getLocationChangeInfo,
+  handleHashScroll,
+  trimPathRight,
+} from '@tanstack/router-core'
 import { useRouter } from './useRouter'
 import { useRouterState } from './useRouterState'
 import { usePrevious } from './utils'
@@ -126,23 +130,7 @@ export function Transitioner() {
             resolvedLocation: s.location,
           }))
 
-          if (
-            typeof document !== 'undefined' &&
-            (document as any).querySelector
-          ) {
-            const hashScrollIntoViewOptions =
-              router.state.location.state.__hashScrollIntoViewOptions ?? true
-
-            if (
-              hashScrollIntoViewOptions &&
-              router.state.location.hash !== ''
-            ) {
-              const el = document.getElementById(router.state.location.hash)
-              if (el) {
-                el.scrollIntoView(hashScrollIntoViewOptions)
-              }
-            }
-          }
+          handleHashScroll(router)
         }
       },
     ),
